refact: use Robolectric at :platform:autos
