--- platform/autos/src/androidTest/java/br/com/orcinus/orca/platform/autos/kit/action/button/skeleton/SemanticsNodeInteractionsProviderExtensionsTests.kt
@@ -1,42 +1,0 @@
-/*
- * Copyright © 2023–2024 Orcinus
- *
- * This program is free software: you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
- * the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program. If
- * not, see https://www.gnu.org/licenses.
- */
-
-package br.com.orcinus.orca.platform.autos.kit.action.button.skeleton
-
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.junit4.createComposeRule
-import br.com.orcinus.orca.platform.autos.kit.action.button.skeleton.test.onLoadingIndicator
-import org.junit.Rule
-import org.junit.Test
-
-internal class SemanticsNodeInteractionsProviderExtensionsTests {
-  @get:Rule val composeRule = createComposeRule()
-
-  @Test
-  fun findsLoadingIndicator() {
-    composeRule.setContent {
-      Button {
-        DisposableEffect(Unit) {
-          load()
-          onDispose {}
-        }
-
-        Loadable {}
-      }
-    }
-    composeRule.onLoadingIndicator().assertIsDisplayed()
-  }
-}
