+++ platform/autos/src/test/java/br/com/orcinus/orca/platform/autos/kit/scaffold/bar/navigation/view/image/WithImageTintMatcher.kt
